+++ src/tezos.py
@@ -155,17 +155,6 @@ async def withdraw(tezos_manager, to, amount):
     return await tezos_manager.queue_operation(sender=to, operation=op)
 
 
-def check_calls_per_month(db, contract_id):
-    max_calls = crud.get_max_calls_per_month_by_contract_address(db, contract_id)
-    # If max_calls is -1 means condition is disabled (NO LIMIT)
-    if max_calls == -1:  # type: ignore
-        return True
-    nb_operations_already_made = crud.get_operations_by_contracts_per_month(
-        db, contract_id
-    )
-    return max_calls >= len(nb_operations_already_made)
-
-
 class TezosManager:
     def __init__(self, ptz):
         self.ops_queue = OrderedDict()
