+++ src/utils.py
@@ -1,3 +1,5 @@
+from src import crud
+
 # -- EXCEPTIONS --
 
 
@@ -39,3 +41,14 @@ class NotEnoughFunds(Exception):
 
 class TooManyCallsForThisMonth(Exception):
     pass
+
+
+def check_calls_per_month(db, contract_id):
+    max_calls = crud.get_max_calls_per_month_by_contract_address(db, contract_id)
+    # If max_calls is -1 means condition is disabled (NO LIMIT)
+    if max_calls == -1:  # type: ignore
+        return True
+    nb_operations_already_made = crud.get_operations_by_contracts_per_month(
+        db, contract_id
+    )
+    return max_calls >= len(nb_operations_already_made)
